Extract OngoingStubbing extensions for Observable to other class
